// src/features/accounting/shared/types/sharedTypes.ts
export type Voucher = {
  id?: number;
  type?: number; // por alguna razon type solo aparece cuando haces PUT, despues es un query string
  num?: number;
  exchangeRate: number;
  voucherDate?: string | Date;
  coin: "USD" | "BOB"; //this should be an enum
  checkNum?: string | undefined;
  canceledTo?: string | Date;
  gloss: string;
  bankId: number | string | null;
  items?: VoucherItem[];
  hojaDeRuta?: string;
  provider?: string;
  nit?: string;
  invoice?: string;
  invoiceNumber?: string;
};

export type VoucherItem = {
  id?: number;
  debitBs: number | null;
  debitSus: number;
  assetBs: number | null;
  assetSus: number;
  gloss: string;
  accountId?: number | string;
  code?: string;
  description?: string;
  hojaDeRuta?: string;
  typeOfExpense?: any;
  createdAt?: "string";
  voucherId?: number | string;
  type?: number;
  canDebit?: boolean //para el uso de inputs en los asientos modelo
  canAsset?: boolean
  percentage?: number // para el calculo de los porcentajes en los asientos modelo
  accountDescription?: string
};

export enum VoucherType {
  DIARY = "0",
  EXPENSE = "1",
  INCOME = "2",
}

export enum VoucherTypeRoute {
  DIARY = "diary",
  EXPENSE = "expenses",
  INCOME = "income",
}

export type StableCallback<T extends (...args: any[]) => any> = T & { __stable?: true };